refactor: move tea.Cmd and tea.Msg declarations to their own files
